Rename month tests
