+++ tests/Month/CompareToTest.php
@@ -8,27 +8,27 @@ use DigitalCraftsman\DateTimeParts\Month;
 use PHPUnit\Framework\TestCase;
 
 /** @coversDefaultClass \DigitalCraftsman\DateTimeParts\Month */
-final class MonthIsBeforeTest extends TestCase
+final class CompareToTest extends TestCase
 {
     /**
      * @test
      *
      * @dataProvider dataProvider
      *
-     * @covers ::isBefore
+     * @covers ::compareTo
      */
-    public function is_before_works(
-        bool $expectedResult,
+    public function compare_to_works(
+        int $expectedResult,
         Month $month,
         Month $comparator,
     ): void {
         // -- Act & Assert
-        self::assertSame($expectedResult, $month->isBefore($comparator));
+        self::assertSame($expectedResult, $month->compareTo($comparator));
     }
 
     /**
      * @return array<string, array{
-     *   0: boolean,
+     *   0: int,
      *   1: Month,
      *   2: Month,
      * }>
@@ -37,22 +37,22 @@ final class MonthIsBeforeTest extends TestCase
     {
         return [
             'previous year' => [
-                false,
+                1,
                 Month::fromString('2022-10'),
                 Month::fromString('2021-10'),
             ],
             'same month' => [
-                false,
+                0,
                 Month::fromString('2022-10'),
                 Month::fromString('2022-10'),
             ],
             'next year' => [
-                true,
+                -1,
                 Month::fromString('2022-10'),
                 Month::fromString('2023-10'),
             ],
             'next month' => [
-                true,
+                -1,
                 Month::fromString('2022-10'),
                 Month::fromString('2022-11'),
             ],
