+++ tests/Month/IsBeforeTest.php
@@ -8,22 +8,22 @@ use DigitalCraftsman\DateTimeParts\Month;
 use PHPUnit\Framework\TestCase;
 
 /** @coversDefaultClass \DigitalCraftsman\DateTimeParts\Month */
-final class MonthIsAfterTest extends TestCase
+final class IsBeforeTest extends TestCase
 {
     /**
      * @test
      *
      * @dataProvider dataProvider
      *
-     * @covers ::isAfter
+     * @covers ::isBefore
      */
-    public function is_after_works(
+    public function is_before_works(
         bool $expectedResult,
         Month $month,
         Month $comparator,
     ): void {
         // -- Act & Assert
-        self::assertSame($expectedResult, $month->isAfter($comparator));
+        self::assertSame($expectedResult, $month->isBefore($comparator));
     }
 
     /**
@@ -38,8 +38,8 @@ final class MonthIsAfterTest extends TestCase
         return [
             'previous year' => [
                 false,
-                Month::fromString('2021-10'),
                 Month::fromString('2022-10'),
+                Month::fromString('2021-10'),
             ],
             'same month' => [
                 false,
@@ -48,13 +48,13 @@ final class MonthIsAfterTest extends TestCase
             ],
             'next year' => [
                 true,
-                Month::fromString('2023-10'),
                 Month::fromString('2022-10'),
+                Month::fromString('2023-10'),
             ],
             'next month' => [
                 true,
-                Month::fromString('2022-11'),
                 Month::fromString('2022-10'),
+                Month::fromString('2022-11'),
             ],
         ];
     }
