+++ tests/Month/ToStringTest.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCraftsman\DateTimeParts\Month;
+
+use DigitalCraftsman\DateTimeParts\Month;
+use PHPUnit\Framework\TestCase;
+
+/** @coversDefaultClass \DigitalCraftsman\DateTimeParts\Month */
+final class ToStringTest extends TestCase
+{
+    /**
+     * @test
+     *
+     * @covers ::__toString
+     */
+    public function to_string_works(): void
+    {
+        // -- Arrange & Act
+        $month = Month::fromString('2022-10');
+
+        // -- Assert
+        self::assertEquals('2022-10', (string) $month);
+    }
+}
